{{ cookiecutter.project_name }}/files/requirements.txt
mqttwrapper==0.0.8
python-dotenv==0.14.0
